Resolve a regression with the patch for missing binaries
A recent change to swift-init made it so it checks if binaries
are installed before trying to run them. If binaries are missing,
we do not instantiate a Server, and thus do not even deal with
that.

Simple and effective, but trouble starts because we want
a perfect behavior about exit code in all cases.

To that end, the incorrect patch adds a redundant "missing
binaries" flag to Manager, sets it in case binaries are missing
in any server type. Then, unfortunately, the decorator makes
a shortcut and does not even run the command, and returns
a nonzero exit code.

This leads to a failure in the most basic case, when you have
a storage only node with no proxy. Since proxy is missing,
the missing_binaries flag gets set in the Manager, and then
nothing else is invoked either. The issue is, if binaries are
missing for a would-be Server, it should not bring down whole
Manager.

The simplest fix would be just do the verify_server() thing,
and nothing else, but then we never return an error even if
specifically trying to run a missing server type (proxy-server).
So, our fix tries to do a little better: we return an error
if only one type was requested, but not if several were.
This way, an alias (like "main") or a wildcard ("*") will
not return an error, if all available services start okay.
But it seems fine, right?

Related-Change-Id: I09143ae20bd6249083c0b80cdaa9d561e6abb301
Change-Id: I6032ac28a785ec3efa4fdcc73ed868c733214bd5