Fix logging in proxy container GET path
The proxy tried to map the x-backend-sharding-state header value to
a ShardRange state, but the value is the container *db* state. The
attempted mapping would commonly fail because db state names do not
always correspond to ShardRange state names, in which the case the
fallback was to log the header value i.e. the correct outcome.

Sometimes the attempted mapping would succeed because for example
'sharded' is both a db state name and a ShardRange state name. In that
case the log message would look something like:

  "Found 1024 objects in shard (state=(70, 'sharded')), total = 1024"

i.e. the tuple of ShardRange state name and number was logged, which
was inappropriate.

Change-Id: Ic08e6e7df7162a4c1283a3ef6e67c3b21a4ce494