sharder: Ignore already shrinking sequence when compacting
If a sequence of shard ranges is already shrinking then in some
circumstances we do not want to report it as a candidate for
shrinking. For backwards compatibility allow already shrinking
sequences to be optionally included in return value of
find_compactible_shard_sequences.

Also refactor to add an is_shrinking_candidate() function.

Change-Id: Ifa20b7c08aba7254185918dfcee69e8206f51cea