sharder: downgrade 'no data dirs' warning to info
Previously if the sharder found no 'containers' directory on any
device then it would emit a warning. This can be a distraction when in
normal operation the sharder daemon is run on all nodes including
those not in the container ring. This patch clarifies the log message
and downgrades it to info level:

  INFO: Found no containers directories

The change also fixes the intermittent failure of the probe test
test_delete_root_reclaim in test/probe/test_sharder.pr. The test
asserts that no warning logs are emitted, but would previously fail if
no containers were created on one of the devices, since the 'Found no
data dirs' warning would be emitted.

Change-Id: I97f7aa790c225ffd72b77a3fc1a6ef41f72df13e