Fix race in diskfile.relink_paths
When a diskfile is written in a new diskfile dir, then in
diskfile.relink_paths the existence of the diskfile dir is tested
before a call to makedirs. If more than one process is taking the same
action concurrently they may race such that one tests for existence,
then another call makedirs, then the first calls makedirs and an
exception is raised. The first process therefore fails to relink the
object file, which is particularly problematic if the first process
was writing a newer object file.

This patch changes relink_paths to unconditionally attempt the
makedirs but catch and ignore any EEXIST errors.

Change-Id: Ia041f7b7dbd4e264072325b4b52718fef9966708