relinker: make cleanup checks more robust
Before removing object files from the old partition ensure that all
required links exist in the new partition.  Allow for the object files
in the old partition to not necessarily be among the required files
if, for example, newer files have been created in the new partition
since relinking occurred.

When required links do not already exist in the new partition an
attempt is made to create them during cleanup.

Co-Authored-By: Tim Burke <tim.burke@gmail.com>
Change-Id: I6e01ed29eb8971eeff96887efbfdfb76a01b00c3