Relinker: Add partition progress to relinker log
Turns out the relinker's relink and cleanup command can take a while so
it would be nice to have progress added to the log file.

This patch adds an info level log line to the relinker log once a
partition is completed. It looks something like:

  Device: sda1 Step: cleanup Partitions: 1/2

So from a glance we can see what device, what step and the progress. In
the above snippet it's just finished the first of 2 partitions.

Change-Id: I364b37ca77a5b6d6125a1f133efa255340316949