Fix systemd notify unit test on macos
MacOS does not support the abstract socket addressing [1] used by
test_systemd_notify, so the test previously failed on MacOS. With this
patch the test using an abstract address is only attempted on linux
platforms. An additional test is added for all platforms that binds
the socket to a filesystem path.

[1] https://man7.org/linux/man-pages/man7/unix.7.html

Change-Id: Ibb77995cc4fcd95bfbfbf0095a4b7f166c436eb4