swift-manage-shard-ranges analyze: accept incomplete shard data
swift-container-info does not currently include 'epoch' among the
shard range info that it outputs, which prevents shard data scraped
from that tool being read by the swift-manage-shard-ranges 'analyze'
command.  The 'epoch' key is not significant to the 'analyze' results
so it is safe to add a default value when it is missing.

Change-Id: I7be30c3247a53969a17fed63d0eba5ab4f4027e4