Add sharding to swift-recon
Note that this does not modify the recon middleware: we already
support the /sharding endpoint in it. But inexplicably there's
no CLI that interrogates and parses the sharding information.

The overview_container_sharding.rst document tells operators
to read raw JSON. It has the advantage of always having the
full picture, but these days we deserve a digest by CLI.

Change-Id: Iac71f68f6633764e0c926ca60990be3b16ef6855