swift-manage-shard-ranges: fix exit codes
Previously swift-manage-shard-ranges would variously return either 1
or 2 in cases of invalid arguments, errors or unexpected outcomes, or
if the user chose to quit before a change was applied. This patch
applies a more consistent pattern to the exit codes:

0 = success
1 = an unexpected outcome, including errors
2 = invalid command line arguments or conf file options
3 = user quit

Some errors that previously resulted in an exit code 2 will now exit
with code 1.

Change-Id: Icf170fef26ed36aab3bf845c5560f1e579a69c2b