Drop the unnecessary sudo from probe tests
While innocently running probe tests that usued to work,
they started dropping to [sudo] prompt and hang. Apparently,
someone added an harmless-looking sudo, but this does not work
if the testing user does not have sudo priviletes.

The code checks very explicitly above that these operations
would be allowed, and thus the sudo is unnecessary. We can
just omit it, and get probe tests working again.

Change-Id: I654c3d954711d4892ff43d7a8df7c3248cdf7aff