relinker: let --workers accept auto
Currently from the relinker cli we allow a --workers parameter but it
currently _must_ be an int. However, if someone wants to override a
static int worker value from the config back to the 'auto' this results
in a VALUE error.

This patch changes the argparse type check for the worker param to use
config_auto_int_value, via a small shim, so a value of 'auto' will be
accepted.

Closes-Bug: #1926757
Change-Id: I5691bf644ade9fd7f7490f91add388a1d5a4aaa2