Do not use epoll in relinker
In all daemons and servers we use the poll hub, there's a note about
epoll locking up when a watched fd gets closed.

We observed the relinker locking up and hung in epoll (perhaps related
to the logging PipeMutex fd getting closed?)

We should use the same hub in the relinker as other swift processes for
consistency.

Change-Id: Ib9a774a07de2fe5fe4577daa7545f430a5764b9b