Sharding: root audit epoch reset warning
We've seen an epoch get reset on a root container in production. I
imagine this could happen if get_own_range is called allowing the self
generated default being called on a new primary before replication has
moved the old db across.

This patch doesn't fix the issue, but adds a root audit reset epoch
check so we can log when/if it happens. So we can 1. see when it happens
and how systemic it is. and 2. if it needs to be fixed automatically or
something that can be done on an ad-hoc basis.

Change-Id: Ia7584ea35600df67dae1cf3694e0ef62612931fa