sharder: If saving own_shard_range use no_default=True
This is just a belts and braces patch that makes sure that we use
own_shard_range(no_default=True) if we assume there will always be an
own_shard_range and it might be written back to the broker.

I'm not sure it's even possible for there not to be an own_shard_range when
we get to `_cleave_shard_range` or `_complete_sharding`. But in any case we
definitely shouldn't allow the possiblity of a default own_shard_range in
any case.

Change-Id: I68c3e7558d04e8dcd5874f80b4763ec0b820331d