Fix docker image builds
The image builds are trying to fetch a gpg pubkey from keybase which
does not exist. This causes the builds to fail. Fix this by vendoring
the key in the swift repo directly. This should be more reliable.

Upstream bug at https://github.com/just-containers/s6-overlay/issues/352

Note I looked up the pubkey given the hash on that issue and copied its
contents into this change. I can't vouch for the validity of this key
beyond that. However, if it is listed in that issue and validates the
packages from s6 then it is probably good.

Change-Id: I3f42fbaae5f246836024f95b549b487b41bd6f1d