Bug: fix s3api multipart parts listings
s3api returns multipart parts listings out of order and possibly
missing. For example, if there are 2000 parts, the first 12 parts
returned by s3api currently will be: 1, 10-19, 100. Then after part
199, the following part is 1000, and so on.

The change fixes this behavior by internally listing all of the parts
(with default settings, this should be 1 listing request, as the 10000
parts limit matches the Swift listing limit). After that, the parts are
sorted and delimited/marker settings are applied to craft the response
for the client.

Change-Id: I150cf53b07e7d2d8de1d6e8c1fb08c07b9afe842